+++ src/elisa/analytics/binary_fit/shared.py
@@ -5,17 +5,16 @@ import numpy as np
 from scipy import interpolate
 from abc import ABCMeta, abstractmethod
 
+from elisa.utils import is_empty
+from elisa.conf import config
 from elisa.analytics.params import parameters
 from elisa.analytics.params.parameters import BinaryInitialParameters
 from elisa.analytics.tools.utils import (
     radialcurves_mean_error,
     lightcurves_mean_error,
     time_layer_resolver)
-from elisa.binary_system.system import BinarySystem
 from elisa.observer.observer import Observer
 from elisa.observer.utils import normalize_light_curve
-from elisa.utils import is_empty
-from elisa.conf import config
 from elisa.binary_system.system import BinarySystem